refactor: Pull out a base class from WhatsIncludedAdapter
CustomizeInnerAdapter and WhatsIncludedAdapter now inherit from abstract base class DrinkComponentBaseAdapter.
